makefile

+++ Sort/radixsort.h
@@ -8,12 +8,12 @@ void countingSort(std::vector<int> &vec, std::vector<int> &D, int max_num)
 {
     std::vector<int> Ctr;
     Ctr.assign(max_num + 1, 0);
-    for (int i = 0; i < vec.size(); i++)
+    for (int i = 0; i < (int)vec.size(); i++)
     {
         Ctr[D[i]]++;
     }
 
-    for (int i = 0; i < Ctr.size() - 1; i++)
+    for (int i = 0; i < (int)Ctr.size() - 1; i++)
     {
         Ctr[i + 1] += Ctr[i];
     }
@@ -33,7 +33,7 @@ void radixSort(std::vector<int> &A, int d, int k = 9)
     for (int i = 0; i < d; i++)
     {
         std::vector<int> digit(A.size(), 0);
-        for (int j = 0; j < A.size(); j++)
+        for (int j = 0; j < (int)A.size(); j++)
         {
             digit[j] = A[j] / (int)std::pow(10, i) % 10;
         }
